+++ dumps/shell/client_id.sh
@@ -4,7 +4,7 @@
 curl \
   --silent \
   -c $COOKIE \
-  -d "$(jq -rc ".[0].id = $ID" json/handshake.json)" \
+  -d "$(jq -rc --arg ID "$ID" '.[0].id = $ID' json/handshake.json)" \
   -H 'Content-Type: application/json' \
   '[::1]:1025/notifications/handshake' |
   jq -r '.[0].clientId'