+++ CHANGELOG.md
@@ -4,11 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
-## [0.4.2]
+## [0.5.0]
 
 ### Fixed
 
 - Check channel before start wait for client message.
+- Support `subscribe` field be single-value or array.
 
 ### Add